+++ src/GitHub_Updater/API/GitHub_API.php
@@ -129,9 +129,10 @@ class GitHub_API extends API implements API_Interface {
 		self::$method       = 'download_link';
 		$download_link_base = $this->get_api_url( '/repos/:owner/:repo/zipball/', true );
 		$endpoint           = '';
+		$switch_to_branch   = $branch_switch && array_key_exists( $branch_switch, $this->type->branches );
 
 		// Release asset.
-		if ( $this->type->release_asset && '0.0.0' !== $this->type->newest_tag ) {
+		if ( $this->type->release_asset && '0.0.0' !== $this->type->newest_tag && ! $switch_to_branch ) {
 			$release_asset = $this->get_release_asset();
 			if ( property_exists( $this->type, 'is_private' ) && $this->type->is_private ) {
 				return $this->get_release_asset_redirect( $release_asset, true );