+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* allow for repos using release assets to have branch switcher, switching away from `master` or tag will not use release asset for updating
 
 #### 9.4.1 / 2020-04-04 **HotFix**
 * fixed problem with sanitizing